+++ kite-core/src/main/kotlin/com/tang/kite/metadata/IndexMeta.kt
@@ -47,6 +47,6 @@ class IndexMeta(
     /**
      * Whether this index represents the primary key
      */
-    val isPrimaryKey: Boolean
+    val primaryKey: Boolean
 
 )