+++ kite-core/src/test/kotlin/com/tang/kite/schema/synchronization/column/initialcolumn/AccountInitialColumn.kt
@@ -1,4 +1,4 @@
-package com.tang.kite.schema.synchronization.table.tablecommentnull
+package com.tang.kite.schema.synchronization.column.initialcolumn
 
 import com.tang.kite.annotation.Table
 import com.tang.kite.annotation.id.Id
@@ -7,14 +7,14 @@ import com.tang.kite.annotation.id.IdType
 /**
  * @author Tang
  */
-@Table(value = "account")
-class AccountTableCommentNull(
+@Table(value = "column_account")
+class AccountInitialColumn(
 
     @Id(type = IdType.AUTO)
     var id: Long? = null,
 
     var username: String? = null,
 
-    var password: String? = null,
+    var password: String? = null
 
 )