+++ kite-core/src/test/kotlin/com/tang/kite/schema/synchronization/index/initialindex/AccountInitialIndex.kt
@@ -1,4 +1,4 @@
-package com.tang.kite.schema.synchronization.column.dropcolumn
+package com.tang.kite.schema.synchronization.index.initialindex
 
 import com.tang.kite.annotation.Table
 import com.tang.kite.annotation.id.Id
@@ -7,13 +7,15 @@ import com.tang.kite.annotation.id.IdType
 /**
  * @author Tang
  */
-@Table(value = "column_account")
-class AccountDropColumn(
+@Table(value = "index_account")
+class AccountInitialIndex(
 
     @Id(type = IdType.AUTO)
     var id: Long? = null,
 
     var username: String? = null,
+
+    var nickname: String? = null,
 
     var password: String? = null