+++ kite-core/src/test/kotlin/com/tang/kite/schema/synchronization/table/initialtable/AccountInitial.kt
@@ -1,4 +1,4 @@
-package com.tang.kite.schema.synchronization.column.dropcolumn
+package com.tang.kite.schema.synchronization.table.initialtable
 
 import com.tang.kite.annotation.Table
 import com.tang.kite.annotation.id.Id
@@ -7,14 +7,14 @@ import com.tang.kite.annotation.id.IdType
 /**
  * @author Tang
  */
-@Table(value = "column_account")
-class AccountDropColumn(
+@Table(value = "table_account", comment = "Table for account initial")
+class AccountInitial(
 
     @Id(type = IdType.AUTO)
     var id: Long? = null,
 
     var username: String? = null,
 
-    var password: String? = null
+    var password: String? = null,
 
 )